% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.lineage.description}
\alias{add.lineage.description}
\title{Add a lineage discription to an OTU}
\usage{
add.lineage.description(OTU, add_OTU = TRUE)
}
\arguments{
\item{OTU}{named vector}

\item{add_OTU}{add the OTU name in brackets to the end of the description.}
}
\description{
Takes a named vector (assumed to be a row from a data frame with columns "OTU", "Kingdom", "Phylum", "Class" etc.) and constructs a human-readable lineage description.
It will try to use the binomial if possible.
Returns the input vector, with an additional "Description" element.
}

