% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_relative_abundance.R
\name{add.relative.abundance}
\alias{add.relative.abundance}
\title{(DEPRECATED) Add relative abundance to an OTU table}
\usage{
add.relative.abundance(...)
}
\description{
(DEPRECATED) Add relative abundance to an OTU table
}
\seealso{
Other add_relative_abundance: \code{\link{add_relative_abundance}}
}

