% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lineage_description.R
\name{add_lineage_description}
\alias{add_lineage_description}
\title{Add lineage descriptions to an OTU table}
\usage{
add_lineage_description(otutable, add_otu = TRUE,
  italicise_binomials = "none")
}
\arguments{
\item{italicise_binomials}{if set to 'markdown' or 'md', will wrap
binomials in asterisks so they appear italicised in markdown output. If
set to 'latex', will wrap binomials in escaped 'textit' tags.}

\item{OTUTable}{tbl with at least an 'OTU' column and at least one
taxonomic rank column}

\item{add_OTU}{add the OTU name in parentheses after the description}
}
\description{
Takes a tbl with columns 'OTU', 'Domain', 'Phylum', 'Class'
etc. and constructs a human-readable lineage description, the binomial if
possible. Returns the tbl, with an additional 'Description' column.
}
