% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_abundance.R
\name{add_relative_abundance}
\alias{add_relative_abundance}
\title{Add relative abundance to an OTU table}
\usage{
add_relative_abundance(OTUTable)
}
\arguments{
\item{OTUTable}{the OTU table in a tbl, with at least 'Sample', 'OTU' and
'Count' columns}
}
\description{
Takes an OTU tbl in tidy format, i.e. with at least 'Sample', 'OTU' and
'Count' columns. Appends a 'RelativeAbundance' column containing the relative
abundance (expressed as a percentage) of each OTU within each sample.
}
\seealso{
Other add_relative_abundance: \code{\link{add.relative.abundance}}
}

