% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.word.R
\name{as.word}
\alias{as.word}
\title{Convert an integer to a word}
\usage{
as.word(Number)
}
\arguments{
\item{Number}{number to be converted}
}
\description{
This is a work in progress and currently only supports integers
1-20 inclusive. Very brittle!
}

