% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.taxon.table.R
\name{collapse.taxa.table}
\alias{collapse.taxa.table}
\title{Synonym for "collapse.taxon.table"}
\usage{
collapse.taxa.table(...)
}
\description{
Synonym for "collapse.taxon.table"; deprecated.
}

