% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collapse.taxon.table}
\alias{collapse.taxon.table}
\title{Collapse a taxon table}
\usage{
collapse.taxon.table(TaxonTable, Rank = "Phylum", n = 8,
  Groups = c("Sample"), MergeMinorUnclassified = TRUE,
  UnclassifiedTerms = c(""))
}
\arguments{
\item{TaxonTable}{taxon table in tidy format, i.e. with columns for "Sample", "RelativeAbundance", and at least one rank. Use "add.relative.abundance" first if the table only contains counts.}

\item{Rank}{the rank to collapse.}

\item{n}{the total number of groups following the collapse, including any "Minor" and "Unclassified" groups.}

\item{Goups}{The grouping factor(s), for example if the group mean is to be taken after collapsing.}

\item{MergeMinorUnclassified}{if set to FALSE, will keep minor and unclassified taxa as separate (but still collapsed) groups.}

\item{UnclassifiedTerms}{taxa matching any of these terms will be considered unclassified.}
}
\description{
Takes a table of taxa, or OTUs with taxonomic information e.g. a "Phylum" column, and collapses the taxa of a specified rank which are not in the top n taxa for that rank.
This is useful for plotting the taxonomic composition of a set of samples, where if would be impractical to plot more than a small number of taxa.
}

