% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_taxon_table.R
\name{collapse_taxon_table}
\alias{collapse.taxon.table}
\alias{collapse_taxon_table}
\title{Collapse a taxon table}
\usage{
collapse_taxon_table(TaxonTable, Rank = "Phylum", n = 8,
  UnclassifiedTerms = c(""))

\method{collapse}{taxon.table}(...)
}
\arguments{
\item{TaxonTable}{taxon table in tidy format, with columns named "Sample",
"RelativeAbundance", and at least one rank. Use "add.relative.abundance"
first if the table only contains counts.}

\item{Rank}{the rank to collapse, e.g. "Phylum".}

\item{n}{the desired number of taxa following the collapse, including any
"Minor" and "Unclassified" groups.}

\item{UnclassifiedTerms}{taxa matching any of these terms will be
considered unclassified.}
}
\description{
Takes a table of taxa, or OTUs with taxonomic information e.g. a "Phylum"
column, and collapses the taxa of a specified rank which are not in the
top n taxa for that rank. This is useful for plotting the taxonomic
composition of a set of samples, where if would be impractical to plot
more than a small number of taxa.

The most abundant taxa are determined with the following process:

1. Relative abundances of taxa at the specified rank are summed within
each sample;
2. The mean relative abundance of each taxon across all samples (including
samples where the relative abundance is 0) is calculated;
3. The top n(ish) taxa, excluding unclassified taxa, are selected.
}

