% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deepest.rank}
\alias{deepest.rank}
\title{Identify the deepest rank for which there is taxonomic information}
\usage{
deepest.rank(OTU, Ranks = c("Kingdom", "Phylum", "Class", "Order", "Family",
  "Genus", "Species"))
}
\arguments{
\item{OTU}{named vector}

\item{Ranks}{vector with names of ranks to look for, in order}
}
\description{
Takes a named vector (assumed to be a row from a data frame with columns "OTU", "Kingdom", "Phylum", "Class" etc.) and returns the name of the deepest rank with taxonomic information.
}

