% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lineage_description.R
\name{deepest_rank}
\alias{deepest_rank}
\title{Get the deepest rank from a list of ranks}
\usage{
deepest_rank(Row)
}
\description{
Get the deepest rank from a list of ranks
}
\keyword{internal}
