% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/do.PCoA.R
\name{do.PCoA}
\alias{do.PCoA}
\title{Perform PCoA on a distance matrix [DEPRECATED]}
\usage{
do.PCoA(DistanceMatrix)
}
\description{
This function is deprecated; use cmdscale() from the vegan package instead.
}

