% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lineage_description.R
\name{lineage_description_from_row}
\alias{lineage_description_from_row}
\title{Add lineage description to a row}
\usage{
lineage_description_from_row(Row, add_otu, italicise_binomials)
}
\keyword{internal}

