% Generated by roxygen2 (4.0.1): do not edit by hand
\name{move.to.end}
\alias{move.to.end}
\title{Move a factor level to the end of the factor}
\usage{
move.to.end(Factor, Level)
}
\arguments{
\item{Factor}{the factor}

\item{Level}{the level to move to the end}
}
\description{
Move a factor level to the end of the factor.
Useful when plotting.
}

