% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{move_to_end}
\alias{move_to_end}
\title{(DEPRECATED) Move a factor level to the end of the factor}
\usage{
move_to_end(...)
}
\description{
`move_to_end` is deprecated in favour of `forcats::fct_relevel`. `forcats`
can be installed as part of the `tidyverse`:

\code{install.packages("tidyverse")}

To use `fct_relevel` in a manner similar to `move_to_end`, take advantage of
\code{after = Inf}, which will move the specified factor to the end:
\preformatted{library(forcats)
fct_relevel(Phyla, "Minor/Unclassified", after = Inf)}
}
