% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.tidy.dt}
\alias{read.tidy.dt}
\title{Load a data table from a file in "tidy" format.}
\usage{
read.tidy.dt(Path)
}
\arguments{
\item{file}{path of the file data will be read from}

\item{...}{further arguments to be passed to "read.csv"}
}
\description{
See http://vita.had.co.nz/papers/tidy-data.pdf for more info on "tidy data".
}
\seealso{
"write.tidy", to write data frames in tidy format

"read.tidy", read.tidy for data frames
}

