% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_lineage_description.R
\name{rowwise_add_lineage_description}
\alias{rowwise_add_lineage_description}
\title{Add lineage description to a row}
\usage{
rowwise_add_lineage_description(Row, add_OTU = TRUE,
  italicise_binomials = "none")
}
\description{
Add lineage description to a row
}
\keyword{internal}

