% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_wilkox.R
\name{theme_wilkox}
\alias{theme_wilkox}
\title{Standardised ggplot2 theme for publications}
\usage{
Plot + theme_wilkox()
Plot + theme_wilkox() \%+replace\% theme(...)
}
\description{
Standardised ggplot2 theme for publications
}

