% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write.pdf.R
\name{write.pdf}
\alias{write.pdf}
\title{Write a dataframe to pdf}
\usage{
write.pdf(Data, ...)
}
\arguments{
\item{Data}{dataframe to write}

\item{...}{further arguments to pass to kable}
}
\description{
Write a data frame to pdf using kable (from knitr) and pandoc. Pandoc
must be installed.
}

