% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_tidy.R
\name{read.tidy}
\alias{read.tidy}
\title{(DEPRECATED) Write a data frame to tsv}
\usage{
read.tidy(...)
}
\description{
write.tidy() is deprecated; use write_tidy() instead
}
\seealso{
Other tidy_files: \code{\link{read.tidy}};
  \code{\link{write_tidy}}
}

