% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_pdf.R
\name{write_pdf}
\alias{write_pdf}
\title{Write a tbl to pdf}
\usage{
write_pdf(Data, File, ...)
}
\arguments{
\item{Data}{tbl to write}

\item{File}{file to write to (must end in .pdf)}

\item{...}{further arguments to pass to kable}
}
\description{
Write a tbl to pdf using kable (from knitr) and pandoc. Pandoc must be
installed. xelatex is used as the latex engine.
}

