% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_tidy.R
\name{write_tidy}
\alias{write_tidy}
\title{Write a tbl to tsv}
\usage{
write_tidy(data, file, ...)
}
\arguments{
\item{data}{tbl to be written}

\item{file}{path of the file data will be read from}

\item{...}{further arguments to be passed to 'write.table'}
}
\description{
'write_tidy' writes a tbl to file in tidy format. It's just a convenience
wrapper for 'write.table' with arguments set for tidy-formatted files.
}
\seealso{
readr::read_tsv

Other tidy_files: \code{\link{read.tidy}}
}

