% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.lineage.description}
\alias{add.lineage.description}
\title{Add a lineage discription to an OTU}
\usage{
add.lineage.description(OTUTable, add_OTU = TRUE,
  italicise_binomials = NULL)
}
\arguments{
\item{OTUTable}{data frame with at least an "OTU" column and at least one taxonomic rank column}

\item{add_OTU}{add the OTU name to the description}

\item{italicise_binomials}{if set to "markdown" or "md", will wrap binomials in asterisks so they appear italicised in markdown output}
}
\description{
Takes a data frame with columns "OTU", "Kingdom", "Phylum", "Class" etc. and constructs a human-readable lineage description.
It will try to use the binomial if possible.
Returns the data frame, with an additional "Description" column.
}

