% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.lineage.description.R
\name{add.lineage.description}
\alias{add.lineage.description}
\title{Add lineage descriptions to an OTU table}
\usage{
add.lineage.description(OTUTable, add_OTU = TRUE,
  italicise_binomials = "none")
}
\arguments{
\item{OTUTable}{data frame with at least an "OTU" column and at least one
taxonomic rank column}

\item{add_OTU}{add the OTU name in parentheses after the description}

\item{italicise_binomials}{if set to "markdown" or "md", will wrap
binomials in asterisks so they appear italicised in markdown output. If
set to "escaped_latex", will wrap binomials in escaped "\textit{}"s.}
}
\description{
Takes a data frame with columns "OTU", "Kingdom", "Phylum",
"Class" etc. and constructs a human-readable lineage description.  It will
try to use the binomial if possible.  Returns the data frame, with an
additional "Description" column.
}

