% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_relative_abundance}
\alias{add_relative_abundance}
\title{(DEPRECATED) Add relative abundance to an OTU table}
\usage{
add_relative_abundance(...)
}
\arguments{
\item{OTUTable}{the OTU table in a tbl, with at least 'Sample', 'OTU' and
'Count' columns}
}
\description{
`add_relative_abundance` is deprecated. Use the `dplyr` package and the following idiom
instead:

\preformatted{library(dplyr)
OTUTable \%>\%
 group_by(Sample) \%>\%
 mutate(RelativeAbundance = (100 * Count) / sum(Count) \%>\%
 ungroup}
}
\seealso{
Other add_relative_abundance: \code{\link{add.relative.abundance}}
}
\concept{add_relative_abundance}
