% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.relative.abundance}
\alias{add.relative.abundance}
\alias{add.relative.abundance.dt}
\title{Add relative abundance to an OTU table}
\usage{
add.relative.abundance(OTUTable)

add.relative.abundance.dt(OTUTable)
}
\arguments{
\item{OTUTable}{the OTU table, which must be a tidy data frame or data table with at least "Sample", "OTU" and "Count" columns}
}
\description{
Takes an OTU table in tidy format, i.e. with at least "Sample", "OTU" and "Count" columns.
Appends a "RelativeAbundance" column containing the relative abundance (expressed as a percentage) of each OTU within each sample.

"add.relative.abundance.dt" is significantly faster if the OTU table is in a data table.
}

