% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.relative.abundance.R
\name{add.relative.abundance}
\alias{add.relative.abundance}
\title{Add relative abundance to an OTU table}
\usage{
add.relative.abundance(OTUTable, silent = TRUE)
}
\arguments{
\item{OTUTable}{the OTU table, which must be a tidy data frame or data
table with at least "Sample", "OTU" and "Count" columns}

\item{silent}{suppress messages (defaults to true)}
}
\description{
Takes an OTU table in tidy format, i.e. with at least "Sample", "OTU" and
"Count" columns. Appends a "RelativeAbundance" column containing the
relative abundance (expressed as a percentage) of each OTU within each
sample.
}

