% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.lineage.description.R
\name{deepest.rank}
\alias{deepest.rank}
\title{Identify the deepest rank for which there is taxonomic information}
\usage{
deepest.rank(OTU, Ranks = c("Kingdom", "Phylum", "Class", "Order", "Family",
  "Genus", "Species"))
}
\arguments{
\item{OTU}{row from a data frame with columns "OTU" and at least one
taxonomic rank}

\item{Ranks}{vector with names of ranks to look for, in order}
}
\description{
Takes a row from a data frame with columns "OTU" and at least
one taxonomic rank, and returns the name of the deepest rank with
taxonomic information.
}

