% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_files.R
\name{read_dist}
\alias{read_dist}
\alias{read_dist_tbl}
\alias{write_dist}
\title{Read and write dist objects to file}
\usage{
read_dist(file, ...)

write_dist(dist, file, ...)

read_dist_tbl(Filename, Row, Column, Value)
}
\arguments{
\item{file}{path for reading or writing.}

\item{...}{further arguments to pass to 'read.csv' or 'write.table'}

\item{dist}{object of type 'dist' to write.}
}
\description{
Read or write 'dist' objects from and to file. read_dist returns a
'dist' object, read_dist_tbl returns a melted tbl.
}
\seealso{
dist
}

