% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.dist}
\alias{read.dist}
\alias{write.dist}
\title{Read and write dist objects to file}
\usage{
read.dist(file, ...)

write.dist(dist, file, ...)
}
\arguments{
\item{file}{path for reading or writing.}

\item{dist}{object of type "dist" to write.}

\item{...}{further arguments to pass to "read.csv" or "write.table"}
}
\description{
Read or write an object of type "dist" from/to file.
}
\seealso{
dist
}

