% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dunn}
\alias{Dunn}
\title{Dunn}
\description{
Data frame of indicator families for various sources.
}
\references{
Dunn, R. R., Fierer, N., Henley, J. B., Leff, J. W., & Menninger, H. L. (2013). Home Life: Factors Structuring the Bacterial Diversity Found within and between Homes. PLoS ONE, 8(5), e64133. doi:10.1371/journal.pone.0064133.s005
}
\keyword{data}

