% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure.fully.crossed.R
\name{ensure.fully.crossed}
\alias{ensure.fully.crossed}
\title{Ensure a tidy data frame is fully crossed}
\usage{
ensure.fully.crossed(...)
}
\description{
This function is deprecated; use expand() from the tidyr package instead.
}

