% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ensure.fully.crossed}
\alias{ensure.fully.crossed}
\title{(DEPRECATED) Ensure a tidy data frame is fully crossed}
\usage{
ensure.fully.crossed(...)
}
\description{
`ensure.fully.crossed` is deprecated; use `tidyr::expand` instead.
}
