% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/move_to_end.R
\name{move_to_end}
\alias{move_to_end}
\title{Move a factor level to the end of the factor}
\usage{
move_to_end(Factor, Level)
}
\arguments{
\item{Factor}{the factor}

\item{Level}{the level to move to the end}
}
\description{
Move a factor level to the end of the factor. Useful when plotting.
}

