% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read.tidy}
\alias{read.tidy}
\title{(DEPRECATED) Read a data frame from tsv}
\usage{
read.tidy(...)
}
\description{
`read.tidy` is deprecated; use `readr::read_tsv` instead
}
\seealso{
Other tidy_files: \code{\link{write.tidy}},
  \code{\link{write_tidy}}
}
\concept{tidy_files}
