% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.tidy}
\alias{read.tidy}
\title{Load a data frame from a file in "tidy" format.}
\usage{
read.tidy(file, ...)
}
\arguments{
\item{file}{the name of the file data will be read from}

\item{...}{further arguments to be passed to "read.csv"}
}
\description{
See http://vita.had.co.nz/papers/tidy-data.pdf for more info on "tidy data".
}

