% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_files.R
\name{read.tidy}
\alias{read.tidy}
\alias{read.tidy.dt}
\alias{write.tidy}
\title{Read and write data in tidy format}
\usage{
read.tidy(file, ...)

read.tidy.dt(file, ...)

write.tidy(data, file, row.names = FALSE, ...)
}
\arguments{
\item{file}{path of the file data will be read from}

\item{...}{further arguments to be passed to "read.csv" or "write.table"}

\item{data}{data frame to be written}

\item{row.names}{passed to write.csv}
}
\description{
"read.tidy" and "read.tidy.dt" read tidy data from a file into a data frame or data table, respectively.

"write.tidy" writes a data frame, data table or matrix to file in tidy format.

These functions are just convenience wrappers for the appropriate base R functions, with arguments set for tidy-formatted files.

See \url{http://vita.had.co.nz/papers/tidy-data.pdf} for more info on tidy data.
}

