% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write.tidy}
\alias{write.tidy}
\title{Write a data frame or matrix to file in "tidy" format.}
\usage{
write.tidy(data, file)
}
\arguments{
\item{data}{data frame to write}

\item{file}{path of the file data will be written to}
}
\description{
See http://vita.had.co.nz/papers/tidy-data.pdf for more info on "tidy data".
}
\seealso{
"read.tidy", to read data frames in tidy format

"read.tidy.dt", read.tidy for data tables
}

