% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{write.tidy}
\alias{write.tidy}
\title{(DEPRECATED) Write a data frame to tsv}
\usage{
write.tidy(...)
}
\description{
`write.tidy` is deprecated; use `readr::write_tsv` instead
}
\seealso{
Other tidy_files: \code{\link{read.tidy}},
  \code{\link{write_tidy}}
}
\concept{tidy_files}
