% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tidy.R
\name{write_tidy}
\alias{write_tidy}
\title{(DEPRECATED) Write a tbl to tsv}
\usage{
write_tidy(...)
}
\description{
write_tidy() is deprecated; use readr::write_tsv() instead
}
\seealso{
readr::write_tsv

Other tidy_files: \code{\link{read.tidy}},
  \code{\link{write.tidy}}
}

